pyproject.toml
[project]
name = "stroptix-lite"
version = "0.1.0"
description = "StrOptix Lite: generate tapered plate combinations from inputs and a plate library, apply simple rules, rank by weight, and export to Excel."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "StrOptix Team" }
]
license = { text = "MIT" }
keywords = ["structural", "optimization", "plates", "steel", "cli", "pandas", "pydantic", "typer"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
]
# Dependencies per PEP 621 must be an array of PEP 508 strings
dependencies = [
  "typer[all]>=0.12.0",
  "pydantic>=2.5,<3.0",
  "pandas>=2.0,<3.0",
  "openpyxl>=3.1,<4.0",
  "XlsxWriter>=3.0,<4.0",
  "PyYAML>=6.0,<7.0",
]

[project.urls]
Homepage = "https://example.com/stroptix-lite"

[project.scripts]
stroptix = "stroptix.cli:app"
stroptix-gui = "stroptix.gui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stroptix*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
